refactor(packer): Clean up structure of rommember and romfile
This also adds a struct-member to each of rommember and romfile which
records the ROM offset of the member / file at the time of sealing the
packer. This data is low-cost, and recording it means that we can
consider a multi-threaded approach to dumping the ROM to disk, if we so
choose.